refactor(styles): fix css files naming
